+++ src/base/repository/student_repo.py
@@ -1,5 +1,5 @@
 from repository.baserepo import BaseRepo
-from base.entity.Student import Student
+from src.base.entity.student import Student
 class Student_Repo(BaseRepo):
 
     def getStudentName(self, id) -> str:
